+++ Dockerfile
@@ -10,11 +10,7 @@ RUN go mod download all
 COPY . /code
 
 RUN mkdir bin
-RUN go build -o ./bin/server ./cmd/server
-
-RUN mkdir bin/pg
-ENV FIRA_EMBEDDED_POSTGRES_BINARIES_PATH=/code/bin/pg
-RUN go run ./cmd/bootstrap
+RUN go build -o ./bin/fira ./cmd/fira
 
 # build the client deps
 FROM node:20-alpine as clientdeps
@@ -56,28 +52,53 @@ RUN yarn workspace @fira/app build
 # final request serving image
 FROM node:20-alpine
 
-RUN apk --no-cache add ca-certificates
+ENV USERNAME=fira
+ENV HOME=/home/lib/fira
+ENV DATA=/var/run/fira
+ENV LANG en_US.utf8
 
-WORKDIR /root/
+RUN set -eux; \
+	addgroup -g 70 -S $USERNAME; \
+	adduser -u 70 -S -D -G $USERNAME -H -h $HOME -s /bin/sh $USERNAME; \
+	mkdir -p $HOME/bin; \
+	chown -R $USERNAME:$USERNAME $HOME
+
+RUN apk --no-cache add ca-certificates su-exec
+
+WORKDIR $HOME
 
 # copy backend resources
-COPY --from=backend /code/bin/server ./
-COPY --from=backend /code/bin/pg ./pg
+COPY --from=backend /code/bin/fira ./bin/
 COPY dist ./dist
 COPY gen ./gen
 
+# set up embedded postgres
+RUN mkdir -p $DATA/pg/data && mkdir -p $DATA/pg/runtime
+RUN chown -R $USERNAME:$USERNAME $DATA && chmod 3777 $DATA
+VOLUME $DATA/pg/data
+
+ENV FIRA_EMBEDDED_POSTGRES_DATA_PATH=$DATA/pg/data
+ENV FIRA_EMBEDDED_POSTGRES_BINARIES_PATH=$DATA/pg/bin
+ENV FIRA_EMBEDDED_POSTGRES_RUNTIME_PATH=$DATA/pg/runtime
+
+RUN su-exec $USERNAME $HOME/bin/fira bootstrap && rm -rf $HOME/.embedded-postgres-go
+
 # copy client resources
-RUN mkdir /root/client
+RUN mkdir $HOME/client
 COPY --from=client /code/workspace/apps/fira-app/public ./client/public
 COPY --from=client /code/workspace/apps/fira-app/package.json ./client/package.json
 COPY --from=client /code/workspace/apps/fira-app/.next ./client/.next
 COPY --from=client /code/workspace/node_modules ./client/node_modules
 COPY ./pg/migrations ./pg/migrations
+COPY ./scripts/entrypoint.sh ./entrypoint.sh
 
 ENV NEXT_TELEMETRY_DISABLED 1
 ENV FIRA_DEBUG=false
-ENV FIRA_CLIENT_DIR=/root/client
-ENV FIRA_MIGRATIONS_DIR=/root/pg/migrations
-ENV FIRA_EMBEDDED_POSTGRES_BINARIES_PATH=/root/pg/bin
+ENV FIRA_CLIENT_DIR=$HOME/client
+ENV FIRA_MIGRATIONS_DIR=$HOME/pg/migrations
 
-CMD ["./server"]
+STOPSIGNAL SIGINT
+EXPOSE 8080
+
+ENTRYPOINT ["/bin/sh", "./entrypoint.sh"]
+CMD ["./bin/fira", "serve"]
