+++ dev.Dockerfile
@@ -1,43 +1,41 @@
 FROM golang:1.20-buster
 
-ARG USERNAME=devuser
-ARG USER_UID=1000
-ARG USER_GID=$USER_UID
-ARG PUID=$USER_UID
-ARG PGID=$USER_GID
+ENV USERNAME=devuser
+ENV HOME=/home/lib/fira
+ENV DATA=/var/run/fira
 
-RUN groupadd --gid $USER_GID $USERNAME
-RUN useradd --uid $USER_UID --gid $USER_GID -m $USERNAME
-RUN groupmod -o -g $PGID "$USERNAME"
-RUN usermod -o -u $PUID "$USERNAME"
+RUN set -eux; \
+	addgroup --gid 70 $USERNAME; \
+	adduser --uid 70 --gid 70 --home $HOME --shell /bin/sh $USERNAME; \
+	mkdir -p $HOME/bin; \
+	chown -R $USERNAME:$USERNAME $HOME
 
 RUN curl -fsSL https://deb.nodesource.com/setup_20.x | bash - && \
     apt-get update && \
     apt-get upgrade -y && \
-    apt-get install -y git make openssh-client nodejs sudo && \
+    apt-get install -y git make openssh-client nodejs sudo gosu && \
     go install github.com/cosmtrek/air@latest
 
-RUN echo "$USERNAME:$USERNAME" | chpasswd -e
-RUN usermod -aG sudo $USERNAME
-RUN echo "$USERNAME ALL=(ALL) NOPASSWD:ALL" >> /etc/sudoers
+RUN mkdir $HOME/go
+RUN mkdir $HOME/app
+RUN mkdir $HOME/npm
+RUN npm config set prefix $HOME/npm
 
-USER $USERNAME
+WORKDIR $HOME/app
 
-RUN mkdir /home/$USERNAME/go
-RUN mkdir /home/$USERNAME/app
-RUN mkdir /home/$USERNAME/npm
-RUN npm config set prefix /home/$USERNAME/npm
+RUN chown -R $USERNAME:$USERNAME $HOME
 
-WORKDIR /home/$USERNAME/app
+ENV PATH="$HOME/npm/bin:$HOME/go/bin:${PATH}"
+ENV NODE_PATH="$HOME/npm/lib/node_modules:${NODE_PATH}"
+ENV GOPATH=$HOME/go
 
-ENV PATH="/home/$USERNAME/npm/bin:/home/$USERNAME/go/bin:${PATH}"
-ENV NODE_PATH="/home/$USERNAME/npm/lib/node_modules:${NODE_PATH}"
-
-ENV GOPATH=/home/$USERNAME/go
 ENV NEXT_TELEMETRY_DISABLED 1
 ENV FIRA_DEBUG=true
-ENV FIRA_CLIENT_DIR=/app/workspace/apps/fira-app
+ENV FIRA_CLIENT_DIR=$HOME/app/workspace
 ENV FIRA_BIND_HTTP=0.0.0.0:8080
+ENV FIRA_EMBEDDED_POSTGRES_DATA_PATH=$DATA/data
+ENV FIRA_EMBEDDED_POSTGRES_BINARIES_PATH=$DATA/bin
+ENV FIRA_EMBEDDED_POSTGRES_RUNTIME_PATH=$DATA/runtime
 
 COPY go.mod go.sum Makefile ./
 
@@ -48,12 +46,11 @@ RUN mkdir workspace
 COPY workspace/package.json workspace/yarn.lock ./workspace/
 COPY workspace/libs ./workspace/libs
 COPY workspace/apps ./workspace/apps
-RUN sudo chown -R $USERNAME:$USERNAME /home/$USERNAME
+RUN chown -R $USERNAME:$USERNAME $HOME
 
 RUN make reqs
 
 COPY . .
 
-ENV USERNAME=$USERNAME
-ENTRYPOINT ["/bin/sh", "scripts/entrypoint.sh"]
-CMD ["./bin/server"]
+# ENTRYPOINT ["/bin/sh", "scripts/entrypoint-dev.sh"]
+CMD ["air"]
